Even Fibonacci numbers
A Fibonacci sequence is a sequence that each number is the sum of two preceding numbers. For the algorithm, I set two values a and b to be equal to 1. The formula for the Fibonacci sequence, a, b = b, a+b. While the value is less than 4 million, add all the even numbers to find the sum of the sequence.