Largest Prime Factor
In order to find the largest prime factor of a number other than itself, first, divide that number by two. If two doesn't work then keep on increasing the number until a divisible number is found. For the algorithm, i is equal to 2 and n is equal to the input number to find the largest prime factor. Then use a while loop; while i is less than or equal to n divide by i. If n is divisible by i, n would then equal to the result of n divide by i. If not add 1 to i. Outside of the loop, return n.