Largest palindrome product
At first, I didn't know what palindrome is. A palindrome is a sequence of characters that you can read the same for both forward and backward. I know how to code to get every three digits number to multiply together, but I didn't know how to check if the number is a palindrome. So I did some research online and I came across this method to check if it is a palindrome. My algorithm starts with a method that would return check and return if the number is a palindrome by using the string method to check if the last three number is the same but backward as the first three number. And then I created two for loop, the second loop is inside the first loop, both in rage from 100-1000. I get each number from each for loop to multiply each other, and then I use the method I wrote at the beginning to check if the number is a palindrome. If it is, I would add the number to the list and print the largest number in that list.