smallest multiple
This problem is to find the smallest number that can be divisible by all the numbers from 1 to 20. My logic was to start from the number 20 and keep increasing to 999999999 and add all the valid number to a list and get out the smallest number from that list. I decided to use 999999999 because I don't know how big that number is going to be. So for my code, I have a list of multiple from 1 to 20, and an empty list for the multiple. I run two for loop. The first one is to loop through all the number from 20 to 999999999, and the second one is inside the first loop and it is to loop through all the numbers from the list of multiples. For each number starting from 20, I get it to divide with each number from the list of multiples, and if it is divisible, I add it to the list. At last, I print out the list. I hit a roadblock is that my code takes too long to run, at it can't finish the code but get an error message that said: "memory error" which simply runs out of memory to run this code. I tried to research other methods to use but I couldn't find any that would fit my code. I tried to put a third number in the first for loop which is 20, so it would execute 20 steps, but the number I get is 20, which I don't think is right. I hope I will be able to fix this code in the future and get it to function correctly.