Sum square difference
In order to do this problem, I will have to find the square of the sum by finding the total of all the number from 1 to 100 and square total, and with that total minus the sum of the square by finding the square of each number from 1 to 100 and add them together. So for my code,  I have two parts. The first one is to find the sum of the square by run a for loop range from 0 to 100. For each number, square that number and add it to the total. The second part is finding the square of the sum. So I run my a for loop from 0 to 100, and I add each number to the total. I find the difference by get the sum of the square minus the square of the sum by get the total from the sum of the square minus the total of square of the sum time the square of the sum. I got the answer is 25164150.